+++ src/Snowflake.ts
@@ -1,31 +1,50 @@
-import * as SDK from 'snowflake-sdk';
+import * as SDK from "snowflake-sdk";
 
-import { ConnectionOptions } from './types/ConnectionOptions';
-import { ExecuteOptions } from './types/ExecuteOptions';
-import { LoggingOptions } from './types/LoggingOptions';
-import { Statement } from './Statement';
+import { ConnectionOptions } from "./types/ConnectionOptions";
+import { ExecuteOptions } from "./types/ExecuteOptions";
+import { LoggingOptions } from "./types/LoggingOptions";
+import { Statement } from "./Statement";
 
 export class Snowflake {
   private readonly sdk_connection;
   private readonly logSql: (sqlText: string) => void;
 
-  /* Creates a new Snowflake instance. */
-  constructor(connectionOptions: ConnectionOptions, loggingOptions: LoggingOptions​​ = {}) {
-    if (loggingOptions.logLevel) { SDK.configure({ logLevel: loggingOptions.logLevel }); }
-    this.logSql = loggingOptions.logSql || null;
+  /*
+   * Creates a new Snowflake instance.
+   *
+   * Set the insecureConnect parameter to true to workaround OCSP errors when connecting.
+   * See: https://github.com/snowflakedb/snowflake-connector-nodejs/issues/16
+   */
+  constructor(
+    connectionOptions: ConnectionOptions,
+    loggingOptions: LoggingOptions = {},
+    insecureConnect = false
+  ) {
+    if (loggingOptions && loggingOptions.logLevel) {
+      SDK.configure({ logLevel: loggingOptions.logLevel });
+    }
+    this.logSql = (loggingOptions && loggingOptions.logSql) || null;
+    if (insecureConnect) {
+      SDK.configure({ insecureConnect: true });
+    }
     this.sdk_connection = SDK.createConnection(connectionOptions);
   }
 
   /** the connection id */
-  get id(): string { return this.sdk_connection.getId(); }
+  get id(): string {
+    return this.sdk_connection.getId();
+  }
 
   /** Establishes a connection if we aren't in a fatal state. */
   connect() {
     return new Promise<void>((resolve, reject) => {
       this.sdk_connection.connect(err => {
-        if (err) { reject(err); }
-        else { resolve(); }
-      })
+        if (err) {
+          reject(err);
+        } else {
+          resolve();
+        }
+      });
     });
   }
 
@@ -36,8 +55,11 @@ export class Snowflake {
   destroy() {
     return new Promise<void>((resolve, reject) => {
       this.sdk_connection.destroy(err => {
-        if (err) { reject(err); }
-        else { resolve(); }
+        if (err) {
+          reject(err);
+        } else {
+          resolve();
+        }
       });
     });
   }
@@ -49,7 +71,7 @@ export class Snowflake {
 
   /** A convenience function to execute a SQL statement and return the resulting rows. */
   execute(sqlText: string, binds?: any[]) {
-    const stmt = this.createStatement({sqlText, binds});
+    const stmt = this.createStatement({ sqlText, binds });
     stmt.execute();
     return stmt.getRows();
   }